Update
Removed the older paint.py that wasn't being used.
Added my name to it and made a note regarding loading a model using Pickle.
It seems to work, the only thing left to do is for Morgan to push her changes.
Then I'll go and hook everything up and test it to see if it can predict.