Changes and Added
Added the model into the script, it can take in arrays and make predictions.
Made it so that the script will remove the image when it's done with it.
Changed the image to output to Grayscale instead of RGBA

Still an issue with the array being morphed after reshaping. So if you write a 1 like normal, it could interpret it as a 7. But if you wrote the 1 sideways like a '/' it'll interpret it as a 1.